requirements.txt
tensorflow==1.13.1
gym==0.13.0
stable-baselines==2.9.0
pyyaml==6.0
tqdm==4.64.0
scikit-learn==0.22